Refactored Conf_matrix so that it bbox compatible; reorganised folder layout of analysis
